Make kickstart rom builder work on MacOSX
This is a pretty big overhaul of the script, that should also benefit
Linux users:

- Drop the use of unadf in favor of xdftool which is already installed
  though amitools as we are also using romtool.
- Don't loopback mount the ISO images but use 7z to extract files. This
  works on non-Linux platforms.
- Don't hard code path to bash (MacOSX built-in bash is oooold)
- Guard unpacking of 3.1.4.1 update. If you're using 3.2 it's not needed
  and should not break the build.

+++ software/scripts/unpack.sh
@@ -1,4 +1,4 @@
-#!/bin/bash
+#! /usr/bin/env bash
 
 BUILD=$PWD/build
 ARCHIVES=$PWD/archives
@@ -17,29 +17,39 @@ unpack_AmigaOS314 A4000T
 cd $BUILD/AmigaOS-3.1.4
 
 printf " * Unzipping AmigaOS 3.1.4.1 Update ..."
-unzip -qqo $ARCHIVES/AmigaOS-3.1.4.1-Update.zip
-if [ -r $ARCHIVES/kick_3.1.4.1.zip ]; then
-  unzip -qqo $ARCHIVES/kick_3.1.4.1.zip
-  rm roms/*.bin
-  mv roms/* ROMs/unsplit_unswapped
-  rm -r roms
+if [ -r  $ARCHIVES/AmigaOS-3.1.4.1-Update.zip ]; then
+  unzip -qqo $ARCHIVES/AmigaOS-3.1.4.1-Update.zip
+  if [ -r $ARCHIVES/kick_3.1.4.1.zip ]; then
+    unzip -qqo $ARCHIVES/kick_3.1.4.1.zip
+    rm roms/*.bin
+    mv roms/* ROMs/unsplit_unswapped
+    rm -r roms
+  fi
+  printf " ok\n"
+  unpack_adfs
+else
+  printf " skipped\n"
 fi
-printf " ok\n"
-unpack_adfs
 cd ..
+
+if which -s 7zz; then
+  SEVEN=7zz
+else if which -s 7z; then
+  SEVEN=7z
+else
+  echo "7-zip not installed."
+  exit 1
+fi
+fi
 
 if [ -r $ARCHIVES/AmigaOS3.2CD.iso ]; then
   mkdir -p $BUILD/AmigaOS-3.2
   printf " * Unpacking AmigaOS 3.2 ..."
-  sudo echo "Need root access"
-  sudo mount $ARCHIVES/AmigaOS3.2CD.iso /mnt -oloop,ro -t iso9660
+
   mkdir -p $BUILD/AmigaOS-3.2/ROMs/unsplit_unswapped
-  cp /mnt/rom/*.rom $BUILD/AmigaOS-3.2/ROMs/unsplit_unswapped
-  cp /mnt/adf/install3.2.adf $BUILD/AmigaOS-3.2/
-  cp /mnt/adf/diskdoctor.adf $BUILD/AmigaOS-3.2/
-  chmod 644 $BUILD/AmigaOS-3.2/ROMs/unsplit_unswapped/*.rom
-  chmod 644 $BUILD/AmigaOS-3.2/*.adf
-  sudo umount /mnt
+  $SEVEN e $ARCHIVES/AmigaOS3.2CD.iso ROM/\* -o$BUILD/AmigaOS-3.2/ROMs/unsplit_unswapped -aoa >/dev/null
+  $SEVEN e $ARCHIVES/AmigaOS3.2CD.iso ADF/Install3.2.adf ADF/DiskDoctor.adf -o$BUILD/AmigaOS-3.2/ -aoa >/dev/null
+  printf "ok\n"
   cd $BUILD/AmigaOS-3.2/
   unpack_adfs
   cd ..
